rename middleware dir
